server/requirements.txt
bcrypt==4.1.3
mysql-connector-python==8.4.0
pyaes==1.6.1
pycryptodome==3.20.0
